+++ Assets/Neoxider/Scripts/Cards/Core/Enums/CardLocation.cs
@@ -12,3 +12,5 @@ namespace Neo.Cards
         Pool
     }
 }
+
+