+++ Assets/Neoxider/Scripts/Cards/Core/Interfaces/ICardContainer.cs
@@ -20,3 +20,5 @@ namespace Neo.Cards
         void Clear();
     }
 }
+
+