+++ Assets/Neoxider/Scripts/Save/Providers/ISaveProvider.cs
@@ -118,3 +118,5 @@ namespace Neo.Save
         void Load();
     }
 }
+
+