+++ Assets/Neoxider/Scripts/Save/Providers/SaveProviderType.cs
@@ -16,3 +16,5 @@ namespace Neo.Save
         File
     }
 }
+
+