+++ Assets/Neoxider/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [5.5.0] - Unreleased
+## [5.5.2] - Unreleased
 
 ### Добавлено